+++ CHANGES.rst
@@ -4,9 +4,15 @@ Changelog
 14.1 (unreleased)
 -----------------
 
+Features
+++++++++
+
 - Fix compatibility with pytest 8.2.
   (`#267 <https://github.com/pytest-dev/pytest-rerunfailures/issues/267>`_)
 
+
+- Add ``--fail-on-flaky`` option to fail the test run with custom exit code
+  when test passed on rerun.
 
 14.0 (2024-03-13)
 -----------------